chore: Clean up project documentation and add new features
- Remove unnecessary documentation files (230+ session notes, reports, logs)
  - Deleted MD files: session summaries, phase reports, fix documentation
  - Deleted TXT files: prompt files, debug logs, analysis reports
  - Deleted SQL files: obsolete migration scripts
  - Deleted 20 LOG files

- Update Claude Code permissions in settings
  - Add all tools to allow list for streamlined workflow

- Add new feature implementations
  - Multi-language support (l10n)
  - Enhanced animations and page transitions
  - Improved responsive layouts and utilities
  - Extended owner dashboard functionality
  - New profile management screens
  - Favorites and saved searches features

- Update existing code for Flutter 3.35.6 compatibility

+++ test/unit/utils/result_test.dart
@@ -135,7 +135,7 @@ void main() {
         final chained = result
             .map((data) => data * 2) // 20
             .map<int>((data) => throw const ValidationException('Error'))
-            .map((data) => (data as int) + 5); // Should not execute
+            .map((data) => data + 5); // Should not execute
 
         expect(chained.isFailure, true);
         expect(chained.exceptionOrNull, isA<ValidationException>());
